+++ src/app/list/[listId]/_components/ListDetailInner/Footer.tsx
@@ -14,6 +14,7 @@ import getBottomSheetOptionList from '@/app/list/[listId]/_components/ListDetail
 import ShareIcon from '/public/icons/share.svg';
 import EtcIcon from '/public/icons/etc.svg';
 import EyeIcon from '/public/icons/eye.svg';
+import Script from 'next/script';
 
 interface BottomSheetOptionsProps {
   key: string;
@@ -35,6 +36,12 @@ interface FooterProps {
   collectCount: number;
 }
 
+declare global {
+  interface Window {
+    Kakao: any;
+  }
+}
+
 function Footer({ data }: { data: FooterProps }) {
   const router = useRouter();
   const path = usePathname();
@@ -42,6 +49,13 @@ function Footer({ data }: { data: FooterProps }) {
   const [isSheetActive, setSheetActive] = useState<boolean>(false);
   const [sheetOptionList, setSheetOptionList] = useState<BottomSheetOptionsProps[]>([]);
   const listUrl = `https://listywave.vercel.app${path}`;
+
+  function kakaoInit() {
+    if (!window.Kakao.isInitialized()) {
+      window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_API_KEY);
+    }
+    console.log('kakaoShareStatus:', window.Kakao.isInitialized());
+  }
 
   const goToCreateList = () => {
     router.push(`/list/create?title=${data.title}&category=${data.category}`);
@@ -78,6 +92,13 @@ function Footer({ data }: { data: FooterProps }) {
 
   return (
     <>
+      <Script
+        src="https://t1.kakaocdn.net/kakao_js_sdk/2.6.0/kakao.min.js"
+        integrity="sha384-6MFdIr0zOira1CHQkedUqJVql0YtcZA1P0nbPrQYJXVJZUkTk/oX4U9GhUIs3/z8"
+        crossOrigin="anonymous"
+        onLoad={kakaoInit}
+        strategy="lazyOnload"
+      />
       {isSheetActive && (
         <ModalPortal>
           <BottomSheet onClose={handleOutsideClick} isActive={isSheetActive} optionList={sheetOptionList} />
