+++ src/components/KakaotalkShare/kakaotalkShare.tsx
@@ -22,18 +22,19 @@ function kakaotalkShare({
   const itemTitle2 = listItem[1]?.title ?? '';
   const itemTitle3 = listItem[2]?.title ?? '';
   let allWriter = '';
+
   if (collaborators) {
-    allWriter = [userNickname, ...collaborators].join(',');
+    const collaboratorsNickname = collaborators.map((collaborator) => collaborator.nickname);
+    allWriter = [userNickname, ...collaboratorsNickname].join(',');
   } else {
     allWriter = userNickname;
   }
-
+  console.log(allWriter);
   window.Kakao.Share.sendCustom({
     templateId: 103935,
     templateArgs: {
       title: title,
       description: description,
-      userNickname: userNickname,
       listId: listId,
       itemTitle1,
       itemTitle2,
